fix parallel open with multiple devices
If more than one aardvark device is connected and several
tools/instances want to open devices it was observed that the
open function fails with IOError:ERR_UNABLE_TO_OPEN. The problem
was that in case the device should be opened with a provided
serial number all devices were opened and the serial was compared.

This is fixed by using the py_aa_find_devices_ext in find_devices.

Note that there is still the possiblity that the error occurs when
devices are added/removed between find_devices() and open().

The find_devices returns now a list of tuples with (port, unique_id).

Signed-off-by: Heiko Thiery <heiko.thiery@gmail.com>